+++ src/cochran_commander.c
@@ -541,7 +541,7 @@ static dc_status_t
 cochran_commander_read_retry (cochran_commander_device_t *device, dc_event_progress_t *progress, unsigned int address, unsigned char data[], unsigned int size)
 {
 	// Save the state of the progress events.
-	unsigned int saved = 0;
+	size_t saved = 0;
 	if (progress) {
 		saved = progress->current;
 	}
